modify: SubmitDetails component created

+++ src/components/SubmitFileItem.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import { FaFileAlt, FaTimes } from 'react-icons/fa';
 
-const SubmitFileItem = ({ file, removeFile }) => {
+const SubmitFileItem = ({ file, removeFile, remove }) => {
   return (
     <Fragment>
       <article className="PodcastItem flex content--start items--center">
@@ -10,7 +10,9 @@ const SubmitFileItem = ({ file, removeFile }) => {
         <h2 className="margin--none PodcastItem__Title text--info">
           <a href={file.url} target="_blank" rel="noreferrer"> { file.name } </a>
         </h2>
-        <FaTimes className="cursor--pointer" onClick={() => removeFile(file)}/>
+        {
+          remove &&  <FaTimes className="cursor--pointer" onClick={() => removeFile(file)}/>
+        }
       </div>
     </article>
     </Fragment>
